\name{RcppStrings-package}
\alias{RcppStrings-package}
\alias{RcppStrings}
\docType{package}
\title{
String manipulation using the boost string algoirthms library.
}
\description{
String manipulation using the boost string algorithms library.
}
\details{
\tabular{ll}{
Package: \tab RcppStrings\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-13\cr
License: \tab GPL (>= v2)\cr
}
}
