\name{trimLeft}
\alias{trimLeft}
\title{Trim whitespace from start of string}
\usage{
  trimLeft(input)
}
\arguments{
  \item{input}{input character vector}
}
\value{
  character vector with leading whitespace removed
}
\description{
  Trim whitespace from start of string
}
\examples{
\dontrun{
trimLeft("\\n\\nString with leading white space")
}
}

