\name{trimRight}
\alias{trimRight}
\title{Trim whitespace from end of string}
\usage{
  trimRight(input)
}
\arguments{
  \item{input}{input character vector}
}
\value{
  character vector with trailing whitespace removed
}
\description{
  Trim whitespace from end of string
}
\examples{
\dontrun{
trimRight("\\n\\nString with trailing white space")
}
}

